requirements.txt
groq
python-dotenv
streamlit
pypdf
python-docx
reportlab
